增加了新的功能函数

+++ algo_plugin/algo_plugin/algo_plugin.h
@@ -32,10 +32,10 @@ public:
 };
 
 typedef bool	(*i_init)();
-typedef void	(*i_dynamic_func_a)(float _arg_input, float &_arg_output);
-typedef void	(*i_dynamic_func_b)(std::deque<float> _arg_input, std::deque<float> &_arg_output);
-typedef void	(*i_dynamic_func_c)(std::deque< std::deque<float> > _arg_input, std::deque<float> &_arg_output);
-typedef void	(*i_dynamic_func_d)(std::deque< std::deque<float> > _arg_input, std::deque< std::deque<float> > &_arg_output);
+typedef void	(*i_dynamic_func_a)(double _arg_input, double &_arg_output);
+typedef void	(*i_dynamic_func_b)(std::deque<double> _arg_input, std::deque<double> &_arg_output);
+typedef void	(*i_dynamic_func_c)(std::deque< std::deque<double> > _arg_input, std::deque<double> &_arg_output);
+typedef void	(*i_dynamic_func_d)(std::deque< std::deque<double> > _arg_input, std::deque< std::deque<double> > &_arg_output);
 
 
 typedef struct  
@@ -53,16 +53,16 @@ bool dynamic_init();
 
 // 这是单参数并且单数据的函数
 // This function has single input argument with single value, the same with output argument.
-void dynamic_func_a(std::string _name,float _arg_input, float &_arg_output);
+void dynamic_func_a(std::string _name,double _arg_input, double &_arg_output);
 
 // 这是一个输入参数且以数组的形式传递进来的函数，亦即一维参数函数
 // This is a function with single input argument and multiple value.
-void dynamic_func_b(std::string _name,std::deque<float> _arg_input, std::deque<float> &_arg_output);
+void dynamic_func_b(std::string _name,std::deque<double> _arg_input, std::deque<double> &_arg_output);
 
 //这是多个输入参数，且每个参数以数组赋值的形式传递进来，即二维数组函数,输出参数为一维数组
 //This function contains two dimensional input and  single dimension output arguments.
-void dynamic_func_c(std::string _name, std::deque< std::deque<float> > _arg_input, std::deque<float> &_arg_output);
+void dynamic_func_c(std::string _name, std::deque< std::deque<double> > _arg_input, std::deque<double> &_arg_output);
 
 //这是多个输入参数，且每个参数以数组赋值的形式传递进来，即二维数组函数
 //This function contains two dimensional input and output arguments.
-void dynamic_func_d(std::string _name, std::deque< std::deque<float> > _arg_input, std::deque< std::deque<float> > &_arg_output);
+void dynamic_func_d(std::string _name, std::deque< std::deque<double> > _arg_input, std::deque< std::deque<double> > &_arg_output);
